+++ common/src/main/java/smartin/miapi/editor/syntax/StandardEditors.java
@@ -1,0 +1,4 @@
+package smartin.miapi.editor.syntax;
+
+public class StandardEditors {
+}